backend/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.27
Flask-SQLAlchemy==3.1.1
marshmallow==3.20.2
Flask-Marshmallow==1.0.0
marshmallow-sqlalchemy==1.0.0
pytest==8.0.2
black==24.1.1
flake8==7.0.0
python-jose==3.3.0
bcrypt==4.1.2
alpha-vantage==2.3.1
requests==2.31.0
Flask-Caching==2.1.0 